+++ FinancialPortal/Models/BankAccount.cs
@@ -9,7 +9,7 @@ namespace FinancialPortal.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double Balance { get; set; }
+        public decimal Balance { get; set; }
         public AccountType Type { get; set; }
 
         public string UserId { get; set; }
